CI: Move mfs read write tests into separate file
